pyproject.toml
[tool.poetry]
name = "fastapi-boilerplate"
version = "0.1.0"
description = ""
authors = ["Tural <tural_m@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.63.0"
uvicorn = "^0.13.4"
asyncpg = "^0.22.0"
gino = "^1.0.1"
gino-starlette = "^0.1.1"
SQLAlchemy-Utils = "^0.36.8"
python-jose = "^3.2.0"
email-validator = "^1.1.2"
alembic = "^1.5.6"
psycopg2 = "^2.8.6"
gunicorn = "^20.0.4"
uvloop = "^0.15.2"
httptools = "^0.1.1"

[tool.poetry.dev-dependencies]
black = {version = "^20.8b1", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
